+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Changed
 - Section control interfaces now include a JSDoc comment with the tab display name and a `{@link}` reference to the containing tab section interface
+- Added `{@link}` comments to `ui.tabs.get()`, `tab.sections.get()`, and `ui.quickForms.get()` overloads
+### Fixed
+- Fixed `ui.getControl()` JSDoc enum `{@link}` rendering
 
 ## [1.0.0] - 2026-04-09
 ### Added